asic refactor
