+++ src/dep.h
@@ -1,0 +1,5 @@
+#include <iostream>
+#include <iomanip>
+#include <string>
+
+using namespace std;
